update uint8_t bug

+++ C_code_vitis/trees_managment.c
@@ -69,7 +69,7 @@ void generate_rando_trees(tree_data trees[N_TREES][N_NODE_AND_LEAFS],
 void mutate_trees(tree_data input_tree[N_TREES][N_NODE_AND_LEAFS], 
                  tree_data output_tree[N_TREES][N_NODE_AND_LEAFS],
                  uint8_t n_features, float mutation_rate, 
-                 uint8_t n_trees, float max_features[N_FEATURE], float min_features[N_FEATURE]) {
+                 uint32_t n_trees, float max_features[N_FEATURE], float min_features[N_FEATURE]) {
 
     uint32_t mutation_threshold = mutation_rate * RAND_MAX;
     uint8_t n_feature;
@@ -101,7 +101,7 @@ void mutate_trees(tree_data input_tree[N_TREES][N_NODE_AND_LEAFS],
 void reproducee_trees(tree_data mother[N_TREES][N_NODE_AND_LEAFS],
                         tree_data father[N_TREES][N_NODE_AND_LEAFS],
                         tree_data son[N_TREES][N_NODE_AND_LEAFS],
-                        uint8_t n_trees){
+                        uint32_t n_trees){
 
 
     for (uint32_t tree_i = 0; tree_i < n_trees && tree_i < N_TREES; tree_i++){
